+++ outlet.py
@@ -1,6 +1,6 @@
 from model import *
 
-def start_outlet(self):
+def start_outlet(self, role):
     self.outlet = tk.Toplevel()
     self.outlet.title("Laundrive")
     self.outlet.geometry("960x540+180+80")
@@ -11,6 +11,7 @@ def start_outlet(self):
     self.canvas.pack(fill="both", expand=True)
 
     database(self)
+    self.role = role
 
     self.canvas.create_text(480, 50, text="Data Outlet", anchor="center", font=("Verdana", 28, "bold"), fill="#b5b3b3")
 
@@ -27,11 +28,12 @@ def start_outlet(self):
     csv_button = create_laporan_button(self, frame=self.outlet, x=60, y=110 ,text="CSV", command=lambda: csv_outlet(self))
     xls_button = create_laporan_button(self, frame=self.outlet, x=160, y=110 ,text="Excel", command=lambda: xls_outlet(self))
 
-    tambah_button = create_crud_button(self, frame=self.outlet, x=360, y=110, text="Tambah Data", command=lambda: tambah_outlet(self))
-    edit_button = create_crud_button(self, frame=self.outlet, x=480, y=110, disabled=len(treeview.selection()) == 0, text="Edit Data", command=lambda: edit_outlet(self))
-    delete_button = create_crud_button(self, frame=self.outlet, x=600, y=110, disabled=len(treeview.selection()) == 0, text="Delete Data", command=lambda: delete_outlet(self))
+    if self.role == "Admin":
+        tambah_button = create_crud_button(self, frame=self.outlet, x=360, y=110, text="Tambah Data", command=lambda: tambah_outlet(self))
+        edit_button = create_crud_button(self, frame=self.outlet, x=480, y=110, disabled=len(treeview.selection()) == 0, text="Edit Data", command=lambda: edit_outlet(self))
+        delete_button = create_crud_button(self, frame=self.outlet, x=600, y=110, disabled=len(treeview.selection()) == 0, text="Delete Data", command=lambda: delete_outlet(self))
 
-    treeview.bind("<ButtonRelease-1>", lambda event: switch([edit_button, delete_button], selection=treeview.selection()))
+        treeview.bind("<ButtonRelease-1>", lambda event: switch([edit_button, delete_button], selection=treeview.selection()))
 
 def csv_outlet(self):
     default_filename = 'data.csv'
