+++ pelanggan.py
@@ -1,6 +1,6 @@
 from model import *
 
-def start_pelanggan(self):
+def start_pelanggan(self, role):
     self.pelanggan = tk.Toplevel()
     self.pelanggan.title("Laundrive")
     self.pelanggan.geometry("960x540+180+80")
@@ -11,6 +11,7 @@ def start_pelanggan(self):
     self.canvas.pack(fill="both", expand=True)
 
     database(self)
+    self.role = role
 
     self.canvas.create_text(480, 50, text="Data Pelanggan", anchor="center", font=("Verdana", 28, "bold"), fill="#b5b3b3")
 
@@ -27,11 +28,12 @@ def start_pelanggan(self):
     csv_button = create_laporan_button(self, frame=self.pelanggan, x=60, y=110 ,text="CSV", command=lambda: csv_pelanggan(self))
     xls_button = create_laporan_button(self, frame=self.pelanggan, x=160, y=110 ,text="Excel", command=lambda: xls_pelanggan(self))
 
-    tambah_button = create_crud_button(self, frame=self.pelanggan, x=360, y=110, text="Tambah Data", command=lambda: tambah_pelanggan(self))
-    edit_button = create_crud_button(self, frame=self.pelanggan, x=480, y=110, disabled=len(treeview.selection()) == 0, text="Edit Data", command=lambda: edit_pelanggan(self))
-    delete_button = create_crud_button(self, frame=self.pelanggan, x=600, y=110, disabled=len(treeview.selection()) == 0, text="Delete Data", command=lambda: delete_pelanggan(self))
+    if self.role != "Owner":
+        tambah_button = create_crud_button(self, frame=self.pelanggan, x=360, y=110, text="Tambah Data", command=lambda: tambah_pelanggan(self))
+        edit_button = create_crud_button(self, frame=self.pelanggan, x=480, y=110, disabled=len(treeview.selection()) == 0, text="Edit Data", command=lambda: edit_pelanggan(self))
+        delete_button = create_crud_button(self, frame=self.pelanggan, x=600, y=110, disabled=len(treeview.selection()) == 0, text="Delete Data", command=lambda: delete_pelanggan(self))
 
-    treeview.bind("<ButtonRelease-1>", lambda event: switch([edit_button, delete_button], selection=treeview.selection()))
+        treeview.bind("<ButtonRelease-1>", lambda event: switch([edit_button, delete_button], selection=treeview.selection()))
 
 def csv_pelanggan(self):
     default_filename = 'data.csv'
